requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
loguru==0.5.3
pydantic==1.7.3
PyJWT==1.7.1
rfc3986==1.4.0
sniffio==1.2.0
stackprinter==0.2.5
starlette==0.13.6
urllib3==1.26.3
uvicorn==0.13.4
